requirements.txt
prefect==2.14.10
boto3==1.34.143
kafka-python==2.0.2
psycopg2-binary==2.9.9
sentence-transformers==2.7.0
pyspark==3.5.1
pyarrow==15.0.2
